backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-allauth>=0.61,<1.0
psycopg2-binary>=2.9,<3.0
celery>=5.3,<6.0
redis>=5.0,<6.0
gunicorn>=21.2,<23.0
python-decouple>=3.8,<4.0
django-filter>=24.0,<25.0
